Added python built to solve readline problems
These build files for python are to address problems with the readline
library implementation as shipped with anaconda. These version link to
the readline library which is in turn linked to ncurses. This removes
the problem of having an extra space at the end of tab completion. For
full details see:

https://github.com/ContinuumIO/anaconda-issues/issues/42

These build files were copied from the following github repo.

https://github.com/kbg/anaconda-rl-fix