BLD : add patch to make sure sip version gets set
enaml uses v2 of all the pyqt apis but these api levels
must be set before any part of pyqt is imported.  If VT loads
before enaml (which it does if you are trying to use
enaml in VT) then pyqt gets imported with some of the api levels
as v1.

Added a patch to the build to make the first thing done in `run.py`
to be setting all the sip api levels to v2

Closes https://github.com/Nikea/Internal/issues/9